Add symlinks to the lilyglyphs and musicexamples packages
If the 'latex' directory is visible to XeLaTeX then these links provide access to lilyglyphs and musicexamples too.
It is assumed that the corresponding repos exist besides the 'base' repository.

I implemented this on Linux.
Mac should behave the same,
but I'm not sure what Windows makes of the symlinks.