(Temporarily) disable Inconsolata as tt font
For some reason xelatex doesn't use the bold version of the Inconsolata font.
This is noticeable for example in headings,
but moreover in lstlistings.

Maybe this has something to do with the fact that Inconsolata doesn't come with a 'regular' style but only with a 'medium'?

As a workaround I just commented out the use of the font in OLLbase (i.e. use the default font).